+++ src/ChoyUtilities/Assets/ChoyUtilities/Common/Runtime/Core/Extensions/FloaterCollection.cs
@@ -15,7 +15,7 @@ namespace ChoyUtilities {
         [BurstCompile]
         public static quaternion RotationFromTransform(this Floater data) {
             if (data.Length < 9) return quaternion.identity;
-            var euler = new Vector3(data[3], data[4], data[5]);
+            var euler = new float3(data[3], data[4], data[5]);
             return quaternion.Euler(euler);
         }
         
@@ -26,7 +26,7 @@ namespace ChoyUtilities {
         
         public static Transform Floater(this Transform obj, Floater data) {
             if (data.Length < 9) return obj;
-            var euler = new Vector3(data[3], data[4], data[5]);
+            var euler = new float3(data[3], data[4], data[5]);
             obj.position = new Vector3(data[0], data[1], data[2]);
             obj.rotation = quaternion.Euler(euler);
             obj.localScale = new Vector3(data[6], data[7], data[8]);