+++ src/ChoyUtilities/Assets/ChoyUtilities/Common/Runtime/Utils/UnmanagedCollection.cs
@@ -1,47 +1,53 @@
 ﻿using System;
-using Unity.Burst;
+using System.Collections.Generic;
 
 namespace ChoyUtilities {
 
     public static partial class HelperCollection {
 
-        [BurstCompile]
         public static int FirstMatchUnmanaged<T>(this T[] set, T value)
-            where T : unmanaged{
-            var index = -1;
+            where T : unmanaged
+        {
             for (var i = 0; i < set.Length; i++) {
-                if (!set[i].Equals(value)) continue;
-                index = i;
-                break;
+                if (EqualityComparer<T>.Default.Equals(set[i], value)) {
+                    return i;
+                }
             }
-            return index;
+            return -1;
         }
-        
-        [BurstCompile]
+
         public static (T[], bool) AddIfNotContainUnmanaged<T>(this T[] set, T value)
-            where T : unmanaged {
-            foreach (var e in set) {
-                if (e.Equals(value)) 
+            where T : unmanaged
+        {
+            foreach (var t in set) {
+                if (EqualityComparer<T>.Default.Equals(t, value)) {
                     return (set, false);
-                break;
+                }
             }
+
             var newSet = new T[set.Length + 1];
-            set.CopyTo(newSet, 0);
-            newSet[^1] = value;
+            Array.Copy(set, newSet, set.Length);
+            newSet[set.Length] = value;
             return (newSet, true);
         }
 
-        [BurstCompile]
-        public static (T[], bool) RemoveIfContainUnmanaged<T>(this T[] set, T value) 
-            where T : unmanaged {
+        public static (T[], bool) RemoveIfContainUnmanaged<T>(this T[] set, T value)
+            where T : unmanaged
+        {
             var removeIndex = set.FirstMatchUnmanaged(value);
-            if (removeIndex < -1) return (set, false);
+
+            if (removeIndex == -1) {
+                return (set, false);
+            }
             var newArray = new T[set.Length - 1];
-            if(removeIndex > 0)
+            if (removeIndex > 0) {
                 Array.Copy(set, 0, newArray, 0, removeIndex);
-            if (removeIndex < set.Length - 1)
+            }
+
+            if (removeIndex < set.Length - 1) {
                 Array.Copy(set, removeIndex + 1, newArray, removeIndex, set.Length - removeIndex - 1);
-            
+            }
+
             return (newArray, true);
         }